Refactored, removed rust tests
