+++ src/y_xml.rs
@@ -1,13 +1,18 @@
 use crate::shared_types::{SubId, TypeWithDoc};
 use crate::y_doc::{WithDoc, YDocInner};
+use lib0::any::Any;
 use pyo3::prelude::*;
 use pyo3::types::{PyDict, PyList};
 use std::cell::RefCell;
+use std::collections::HashMap;
 use std::mem::ManuallyDrop;
 use std::ops::Deref;
 use std::rc::Rc;
+use yrs::block::ItemContent;
 use yrs::types::xml::{TreeWalker, Xml, XmlEvent, XmlTextEvent};
+use yrs::types::{BranchPtr, ToJson, TYPE_REFS_MAP, TYPE_REFS_XML_TEXT};
 use yrs::types::{DeepObservable, EntryChange, Path, PathSegment};
+use yrs::MapRef;
 use yrs::XmlFragmentRef;
 use yrs::XmlTextRef;
 use yrs::{GetString, XmlElementPrelim, XmlElementRef, XmlTextPrelim};
@@ -16,6 +21,122 @@ use yrs::{Observable, SubscriptionId, Text, TransactionMut, XmlFragment, XmlNode
 use crate::shared_types::{DeepSubscription, ShallowSubscription};
 use crate::type_conversions::{events_into_py, ToPython, WithDocToPython};
 use crate::y_transaction::{YTransaction, YTransactionInner};
+
+pub fn process_xml_text_node(txn: &TransactionMut<'static>, xml_text_ref: &XmlTextRef) -> Any {
+    let mut result: HashMap<String, Any> = HashMap::new();
+    // Update attributes of the current Text XmlNode
+    let xml_text_map_ref: MapRef = xml_text_ref.clone().into();
+    if let Any::Map(at) = xml_text_map_ref.to_json(txn) {
+        for (k, v) in at.iter() {
+            result.insert(k.to_string(), v.clone());
+        }
+    }
+    if let Some(xml_text_children) = xml_text_ref.successors() {
+        let mut children: Vec<Any> = vec![];
+        let mut child_result: HashMap<String, Any> = HashMap::new();
+        /* xml_text_children contains a sequence of ItemContent instances:
+           ItemContent::Type(YMap) => {"__type": "text", "__format": 0, "__style": "", "__mode": 0, "__detail": 0}
+           ItemContent::String(SplittableString) => "a"
+           ItemContent::String(SplittableString) => " "
+           ...
+           ItemContent::Type(YMap) => {"__type": "text", "__format": 0, "__style": "", "__mode": 0, "__detail": 0}
+           ItemContent::String(SplittableString) => "b"
+        */
+        for child in xml_text_children {
+            match &child {
+                ItemContent::Type(c) => {
+                    let ptr = BranchPtr::from(c);
+                    match ptr.type_ref() {
+                        TYPE_REFS_MAP => {
+                            if !child_result.is_empty() {
+                                children.push(Any::Map(Box::new(child_result)));
+                                child_result = HashMap::new();
+                            }
+                            if let Any::Map(at) = MapRef::from(ptr).to_json(txn) {
+                                for (k, v) in at.iter() {
+                                    child_result.insert(k.to_string(), v.clone());
+                                }
+                            }
+                        }
+                        TYPE_REFS_XML_TEXT => {
+                            let child_xml_text_ref = XmlTextRef::from(ptr);
+                            if !child_result.is_empty() {
+                                children.push(Any::Map(Box::new(child_result)));
+                                child_result = HashMap::new();
+                            }
+                            children.push(process_xml_text_node(txn, &child_xml_text_ref));
+                        }
+                        _ => {
+                            panic!("Unexpected type ref: {:?}", ptr.type_ref());
+                        }
+                    }
+                }
+                ItemContent::String(child_text_part) => {
+                    if !child_result.is_empty() {
+                        let mut child_text = child_result
+                            .get("text")
+                            .unwrap_or(&Any::String("".to_string().into()))
+                            .to_string();
+                        child_text.push_str(child_text_part.as_str());
+                        child_result.insert("text".to_string(), Any::String(child_text.into()));
+                    }
+                }
+                ItemContent::Deleted(_) => (),
+                _ => {
+                    eprintln!("Ignored child of XmlTextRef: {:?}", child);
+                }
+            }
+        }
+        if !child_result.is_empty() {
+            children.push(Any::Map(Box::new(child_result)));
+        }
+        if !children.is_empty() {
+            result.insert(
+                "children".to_string(),
+                Any::Array(children.into_boxed_slice()),
+            );
+        }
+    }
+    Any::Map(Box::new(result))
+}
+
+pub fn process_xml_node(
+    txn: &TransactionMut<'static>,
+    result: &mut HashMap<String, Any>,
+    first_child_maybe: Option<XmlNode>,
+) {
+    let first_child = match first_child_maybe {
+        Some(first_child) => first_child,
+        None => {
+            return;
+        }
+    };
+    let mut children: Vec<Any> = vec![];
+    match first_child {
+        XmlNode::Text(text) => {
+            children.push(process_xml_text_node(txn, &text));
+            text.siblings(txn)
+                .for_each(|sibling: XmlNode| match sibling {
+                    XmlNode::Text(text) => {
+                        children.push(process_xml_text_node(txn, &text));
+                    }
+                    _ => {
+                        eprintln!("Unhandled XmlNode::Text sibling: {:?}", sibling);
+                    }
+                });
+        }
+        XmlNode::Fragment(fragment) => {
+            eprintln!("Unhandled Fragment: {:?}", fragment);
+        }
+        XmlNode::Element(element) => {
+            eprintln!("Unhandled Element: {:?}", element);
+        }
+    }
+    result.insert(
+        "children".to_string(),
+        Any::Array(children.into_boxed_slice()),
+    );
+}
 
 /// XML element data type. It represents an XML node, which can contain key-value attributes
 /// (interpreted as strings) as well as other nested XML elements or rich text (represented by
@@ -575,6 +696,28 @@ impl YXmlFragment {
     /// Returns a string representation of this XML node.
     pub fn __str__(&self) -> String {
         self.0.with_transaction(|txn| self.0.get_string(txn))
+    }
+
+    /// Converts contents of this `YXmlFragment` instance into a Dict representation.
+    pub fn to_dict(&self) -> PyObject {
+        Python::with_gil(|py| {
+            self.0.with_transaction(|txn| {
+                let mut result: HashMap<String, Any> = HashMap::new();
+
+                // Update attributes of the root XmlFragment
+                let xml_fragment_root_map: MapRef = self.0.inner.clone().into();
+                if let Any::Map(at) = xml_fragment_root_map.to_json(txn.deref()) {
+                    for (k, v) in at.iter() {
+                        result.insert(k.to_string(), v.clone());
+                    }
+                }
+
+                // Process the children of the root XmlFragment
+                process_xml_node(txn.deref(), &mut result, self.0.inner.first_child());
+
+                result.into_py(py)
+            })
+        })
     }
 
     /// Returns an iterator that enables a deep traversal of this XML node - starting from first
