+++ y_py.pyi
@@ -235,11 +235,6 @@ def apply_update(doc: YDoc, diff: Union[YDocUpdate, List[int]]):
         apply_update(local_doc, remote_delta)
     """
 
-def update_to_nodes(diff: Union[YDocUpdate, List[int]]) -> Dict:
-    """
-    Applies delta update generated by the remote document to a new document and extracts the nodes
-    from the doc.
-    """
 
 class YTransaction:
     """
@@ -1082,6 +1077,10 @@ class YXmlFragment:
         """
         Returns:
             A string representation of this YXmlFragment
+        """
+    def to_dict(self) -> Dict[str, Any]:
+        """
+        Converts contents of this `YXmlFragment` instance into a Dict representation.
         """
     def get(self, index: int) -> Union[YXmlText, YXmlElement]:
         """
